major bug fixes

+++ litepool/rltrader/litepoolsb3.py
@@ -6,7 +6,7 @@ import torch as th
 import torch.nn as nn
 from packaging import version
 from sb3_contrib import RecurrentPPO
-#from stable_baselines3 import SAC
+from stable_baselines3 import SAC, PPO
 from stable_baselines3.common.policies import ActorCriticPolicy
 from stable_baselines3.common.env_util import make_vec_env
 from stable_baselines3.common.evaluation import evaluate_policy
@@ -30,11 +30,11 @@ from gymnasium import spaces
 class LSTMFeatureExtractor(BaseFeaturesExtractor):
     def __init__(self, observation_space: spaces.Box, lstm_hidden_size: int = 16, output_size: int = 32):
         # The feature dimension is the final output size of the sequential layer
-        super(LSTMFeatureExtractor, self).__init__(observation_space, features_dim=output_size+lstm_hidden_size)
+        super(LSTMFeatureExtractor, self).__init__(observation_space, features_dim=output_size+lstm_hidden_size+24)
         
         self.lstm_hidden_size = lstm_hidden_size
         self.n_input_channels = 30
-        self.remaining_input_size = 14 
+        self.remaining_input_size = 24 
         self.output_size = output_size
         
         self.lstm = nn.LSTM(self.n_input_channels, lstm_hidden_size, batch_first=True)
@@ -69,7 +69,7 @@ class LSTMFeatureExtractor(BaseFeaturesExtractor):
         lstm_hidden_state = lstm_out[:, -1, :]
         
         final_output = self.fc(remaining_input)
-        combined_output = torch.cat((lstm_hidden_state, final_output), dim=1)
+        combined_output = torch.cat((lstm_hidden_state, final_output, remaining_input), dim=1)
         return combined_output 
 
 class VecAdapter(VecEnvWrapper):
@@ -88,18 +88,18 @@ class VecAdapter(VecEnvWrapper):
     self.header = True
 
   def step_async(self, actions: np.ndarray) -> None:
-    self.actions = actions
+      self.actions = actions 
+      self.venv.send(self.actions)
 
   def reset(self) -> VecEnvObs:
       self.steps = 0
       return self.venv.reset()[0]
 
   def seed(self, seed: Optional[int] = None) -> None:
-    pass
+     if seed is not None:
+          self.venv.seed(seed) 
 
   def step_wait(self) -> VecEnvStepReturn:
-      self.venv.send(self.actions)
-    
       obs, rewards, terms, truncs, info_dict = self.venv.recv()
       dones = terms + truncs
       infos = []
@@ -154,49 +154,51 @@ class VecAdapter(VecEnvWrapper):
           print("NaN in OBS...................")
       return obs, rewards, dones, infos
 
-
+import os
+if os.path.exists('temp.csv'):
+    os.remove('temp.csv')
 env = litepool.make("RlTrader-v0", env_type="gymnasium", 
-                          num_envs=4, batch_size=4, 
-                          num_threads=4,
+                          num_envs=1, batch_size=1, 
+                          num_threads=1,
                           filename="deribit.csv", 
                           balance=0.02,
                           depth=20)
 env.spec.id = 'RlTrader-v0'
 env = VecAdapter(env)
 env = VecMonitor(env)
-
+print(env.action_space)
 kwargs = dict(use_sde=True, sde_sample_freq=4)
 
 policy_kwargs = {
     'features_extractor_class': LSTMFeatureExtractor,
     'features_extractor_kwargs': {
         'lstm_hidden_size': 8,
-        'output_size': 16
+        'output_size': 4
     },
     'activation_fn': th.nn.ReLU,
     'net_arch': dict(pi=[16, 32, 8], vf=[16, 32, 16, 8], qf=[16, 32, 16, 8])
 }
 
 import os
-os.remove("temp.csv")
-
 
 model = RecurrentPPO(
-  "MlpLstmPolicy", #CustomGRUPolicy,
+  "MlpLstmPolicy", 
   env,
   policy_kwargs=policy_kwargs,
-  n_steps=4800,
-  learning_rate=1e-3,
-  gae_lambda=0.95,
-  gamma=0.9999,
-  ent_coef=0.01,
-  target_kl=0.06,
+  learning_rate=1e-4,
+  n_steps=256, 
+  batch_size=64,
+  gamma=0.99,
   clip_range=0.1,
+  ent_coef=0.02,
+  vf_coef=0.5,
+  max_grad_norm=0.5,
+  target_kl=0.1,
+  gae_lambda=0.95,
   verbose=1,
   seed=10,
   **kwargs
 )
-
 '''
 model = SAC(
     "MlpPolicy",
@@ -204,10 +206,10 @@ model = SAC(
     policy_kwargs=policy_kwargs,
     learning_rate=3e-4,
     buffer_size=1000000,
-    learning_starts=1000,
-    batch_size=64,
+    learning_starts=200,
+    batch_size=4096*8,
     tau=0.005,
-    gamma=0.99999,
+    gamma=0.99,
     train_freq=1,
     gradient_steps=1,
     ent_coef='auto',
@@ -216,5 +218,4 @@ model = SAC(
     verbose=1,
 )
 '''
-
 model.learn(200000000)
